+++ src/lib/components/Folder.svelte
@@ -1,35 +1,11 @@
 <script>
-	import { getContext } from 'svelte';
-	import { get_opts } from '../utils.js';
+	import Items from '../Items.svelte';
 
 	export let children;
 
-	const { run } = getContext('knobby');
+	// TODO collapsed, label, etc
 </script>
 
 <div class="folder">
-	{#each Object.entries(children) as [name, state]}
-		{#if (state.$visible ? run(state.$visible) : true)}
-			{#if state.component}
-				<div class="item">
-					<span>{name}</span>
-					<div>
-						<svelte:component this={state.component} bind:value={children[name].value} {...get_opts(state)}/>
-					</div>
-				</div>
-			{:else}
-				<button on:click={() => run(state.value)}>{name}</button>
-			{/if}
-		{/if}
-	{/each}
+	<Items bind:children/>
 </div>
-
-<style>
-	.item {
-		display: grid;
-		grid-template-columns: 1fr 200px;
-		grid-gap: 0.5em;
-		align-items: center;
-		min-height: 2em;
-	}
-</style>
